hipify version 2 (#76)
These changes correspond to
https://github.com/pytorch/pytorch/pull/151845 (and syncs changes from
pytorch's hipify as part of this PR). torch's hipify version 2
significantly simplifies the the cuda to hip mappings, including a
BC-breaking change of no longer using caffe2 mappings and getting rid of
the Masquerading files. In order to ease the transition, hipify_torch
will contain both versions of the mappings. An attempt is made to check
which version of torch is installed and use the right mappings,
otherwise the v1 mappings are used unless the caller specifies --v2.